added buildStyle util function, for reuse
