+++ src/DropdownMenu/Popover.svelte
@@ -1,4 +1,6 @@
 <script>
+  import buildStyle from "../utils/buildStyle"
+  
   export let anchor
   export let align = "right"
 
@@ -21,21 +23,6 @@
     if (open && e.key === "Escape") {
       hide()
     }
-  }
-
-  function buildStyle(styles) {
-    const convertCamel = str => {
-      return str.replace(/[A-Z]/g, match => `-${match.toLowerCase()}`)
-    }
-
-    let str = ""
-    for (let s in styles) {
-      if (styles[s]) {
-        let key = convertCamel(s)
-        str += `${key}: ${styles[s]}; `
-      }
-    }
-    return str
   }
 
   function getDimensions() {
